criação de diretórios e subdivisão do código em módulos
