--- README.md
@@ -1,2 +1,0 @@
-# GraficosEnade
-Projeto que gera graficos a partir dos microdados fornecidos pelo Enade.
