+++ bin/preProcessamentoBCC.py
