+++ data/dicionario.txt
